pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "rag-aether"
version = "0.1.0"
description = "RAG system for Aether"
authors = ["Your Name <your.email@example.com>"]
packages = [
    { include = "rag_aether", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.8"
transformers = "^4.48.0"
torch = "^2.5.1"
numpy = "^1.24.0"
faiss-cpu = "^1.7.4"
fastapi = "^0.104.0"
pydantic = "^2.4.2"
python-dotenv = "^1.0.0"
rank-bm25 = "^0.2.2"
sentence-transformers = "^2.2.2"
sentencepiece = "^0.2.0"
uvicorn = "^0.24.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-cov = "^6.0.0"
pytest-xdist = "^3.6.1"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = "-v --cov=src --cov-report=html" 